preprocessing/requirements.txt
torch
torchvision
torchaudio
numpy==1.26.4
pandas
huggingface
wandb
huggingface_hub
transformers
albumentations
scikit-learn
pillow
timm
matplotlib
seaborn
ipython
hydra-core
tqdm
h5py
tensorboard
opencv-python
scikit-image
Pyrebase4
albumentations==0.4.6
